+++ src/modules/webAppMode.mjs
@@ -1,0 +1,2 @@
+// Makes NumÉcoDiag able to be served as webApp to make debugging easier (server required)
+export const webAppMode = false;
